# legacy/architecture/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "ThinkAlike API (Draft Canonical)",
        "version": "0.2.0",
        "description": "Draft OpenAPI specification for the ThinkAlike project. This file defines the canonical schemas for all major entities. Endpoints to be added as implemented."
    },
    "paths": {
...
        "/value-profiles": {
            "post": {
                "summary": "Create a value profile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValueProfile"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Value profile created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValueProfile"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "List all value profiles",
                "responses": {
                    "200": {
                        "description": "A list of value profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValueProfile"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/value-profiles/{profileId}": {
            "get": {
                "summary": "Get a value profile by ID",
                "parameters": [
                    {
                        "name": "profileId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Value profile found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValueProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Value profile not found"
                    }
                }
            },
            "put": {
                "summary": "Update a value profile by ID",
                "parameters": [
                    {
                        "name": "profileId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValueProfile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Value profile updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValueProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Value profile not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a value profile by ID",
                "parameters": [
                    {
                        "name": "profileId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Value profile deleted"
                    },
                    "404": {
                        "description": "Value profile not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValueProfile": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "values": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "archetypes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "resonance_score": {
                        "type": "number"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "user_id",
                    "values",
                    "archetypes",
                    "updated_at"
                ]
            },
...
        }
    }
}